// services/auth-service/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BLOCKED: 'BLOCKED',
  PENDING: 'PENDING'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const TwoFactorMethod = {
  TOTP: 'TOTP',
  EMAIL: 'EMAIL',
  SMS: 'SMS'
} as const

export type TwoFactorMethod = (typeof TwoFactorMethod)[keyof typeof TwoFactorMethod]